Added logic to default to previous level if not all items are complete
